backend/requirements.txt
fastapi
uvicorn
pandas
openai
python-dotenv
matplotlib
plotly
kaleido  # plotly export images library
